ogma-core: Add function to find Variable DB message/port type based on data carried. Refs #219.
Currently, the monitor notifications to the rest of the system happen in
one of two ways: as a call to a user-defined function, or as a
notification in the message bus or via a port. None of those messages
carry any additional information: the occurrence of a property or
requirement violation is implicit in the fact that the notification is
present.

Prior commits have modified several spec definitions to that additional
data (i.e., a type and an expression) can be associated with a
requirement. In order to use that information in the backends, we need
to provide a mechanism to find the type associated with a notification
type (e.g., message, port) even if no input variable is associated to
such type.

This commit modifies the Variable DB so that it provides an additional
module to find the origin type to be used to communicate a result.