ogma-core: Adjust ROS template to carry additional data in monitor handlers. Refs #219.
Currently, the monitor notifications to the rest of the system happen in
one of two ways: as a call to a user-defined function, or as a
notification in the message bus or via a port. None of those messages
carry any additional information: the occurrence of a property or
requirement violation is implicit in the fact that the notification is
present.

Prior commits have extended the ROS backend so the type of any
additional data to be passed by the monitoring system to the handlers is
made available to the template.

This commit modifies the ROS template to use that data to define the
handlers appropriately, and to pass the data in the result.