ogma-cli: Adjust ROS example to demonstrate communicating monitoring results. Refs #219.
Currently, the monitor notifications to the rest of the system happen in
one of two ways: as a call to a user-defined function, or as a
notification in the message bus or via a port. None of those messages
carry any additional information: the occurrence of a property or
requirement violation is implicit in the fact that the notification is
present.

Prior commits have modified the ROS backend so that additional data can
be passed to the handlers when a property being monitored becomes true.

This commit modifies the file formats for the ROS example so that they
demonstrate the new feature.