ogma-core: Adjust known data formats to new spec definitions. Refs #219.
Currently, the monitor notifications to the rest of the system happen in
one of two ways: as a call to a user-defined function, or as a
notification in the message bus or via a port. None of those messages
carry any additional information: the occurrence of a property or
requirement violation is implicit in the fact that the notification is
present.

Prior commits have extended the definition of a spec so that it is
possible to include an additional type an expression for the data to be
communicate to the handlers by the monitors, and have adjusted the JSON-
and XML-based file format parsers accordingly.

This commit modifies all JSON- and XML-based file formats known by Ogma
meet the new definition of a JSONSpec or XMLSpec.