ogma-cli: Include new data available to templates and remove limitations in README. Refs #219.
Currently, the monitor notifications to the rest of the system happen in
one of two ways: as a call to a user-defined function, or as a
notification in the message bus or via a port. None of those messages
carry any additional information: the occurrence of a property or
requirement violation is implicit in the fact that the notification is
present.

Prior commits have modified the ROS, cFS, FPrime and standalone backends
so that additional data can be passed to the handlers when a property
being monitored becomes true.

This commit updates the README to explain the new data available to the
templates, and to remove some of the limitations listed with each
backend.