ogma-core: Add support to read properties from CSV files. Refs #261.
To facilitate importing requirements and properties from different
tools, there is a need to extend Ogma to process CSV files, which are
frequently available as an export option in many of the MBSE tools used
in real-world projects.

A prior commit introduced a library capable of reading CSV files.

This commit modifies `ogma-core` to leverage that library to read
properties or requirements from CSV files.