ogma-core: Add support to read properties from XLSX files. Refs #263.
To facilitate importing requirements and properties from different
tools, there is a need to extend Ogma to process XLSX files, which are
frequently available as an export option in many MBSE tools and used in
many real-world projects.

A prior commit introduced a library capable of reading XLSX files.

This commit modifies `ogma-core` to leverage that library to read
properties or requirements from XLSX files.