ogma-core: List all auxiliary test files in Cabal package. Refs #258.
The Cabal file for ogma-core is not including all files needed for the
tests, or declaring them in the Cabal package, causing the tests to fail
when the packages are unpacked from the distributable source packages
created with Cabal.

This commit changes the Cabal package to include all auxiliary test
files to the extra-source-files field, making sure that the files are
included in the distributable source package but not installed with the
binary distribution.