ogma-cli: Expose optional expression argument to F Prime backend. Refs #121.
To simplify the process of interactive with Ogma, especially when
running quick demos, we'd like to have the possibility of passing an
expression as an argument in the CLI.

A prior commit has extended the F Prime backend to accept an additional
expression as argument and use it as a preferred alternative to the
properties passed in a file.

This commit extends the CLI for the F Prime backend to accept an
expression for that argument via the command line.