ogma-core: Adjust tests for standalone backend. Refs #121.
To simplify the process of interactive with Ogma, especially when
running quick demos, we'd like to have the possibility of passing an
expression as an argument in the CLI.

A prior commit has extended the standalone backend to accept an
additional expression as argument and use it as a preferred alternative
to the properties passed in a file.

This commit adjust the tests of the standalone backend to provide the
right arguments to the backend. Specifically, an additional optional
expression is now provided and the input file is now optional.