ogma-cli: Bump upper version constraint on optparse-applicative. Refs #279.
In order to keep Ogma effectively working in the current Haskell
ecosystem, as well as new versions of GHC, we need to extend the
versions of dependencies that Ogma can be installed with. For example,
optparse-applicative has seen a 0.19.0.0 release, but ogma-cli needs
versions strictly lower than 0.19.

This commit bumps the version constraints on optparse-applicative so
that Ogma can be installed with versions of optparse-applicative lower
than 0.20.