ogma-core: Add to ROS 2 template handling methods for triggers with no args. Refs #287.
The current ROS 2 template does not generate the class methods that
notify of violations when such methods receive no arguments. This
prevents the resulting class from compiling, since the global handlers
cannot find the matching handlers inside the class.

This commit modifies the ROS 2 template to add to the class definition
the handling methods, even when the type of the argument to the handler
is empty (no args).