ogma-core: Install packages locally in ROS 2 dockerfile. Refs #288.
The ROS 2 template installs packages under /opt/spaceros/install, which
is not a good placement for files compiled by the spaceros-user. The
container would be easier to use if files were placed under the user's
$HOME directory.

This commit modifies the installation step for the generated ROS 2
packages in the Dockerfile that accompanies the ROS 2 template so that
the packages are installed locally.