ogma-cli: Add example files to generate ROS 2 monitoring application. Refs #289.
To facilitate adoption of Ogma, we want to have a collection of examples
that demonstrate the most common use cases. A demonstration of how to
use the ROS 2 backend to generate a minimal application would be a good
starting point for users that want to generate monitors for ROS 2.

A prior commit added a guide explaining how to use the ROS 2 backend to
generate a simple monitoring application.

This commit files that can be used with Ogma to generate a simple
monitoring application. The examples match the explanation in the
aforementioned guide.