ogma-cli: Add guide explaining basics of ROS 2 backend. Refs #289.
To facilitate adoption of Ogma, we want to have a collection of examples
that demonstrate the most common use cases. A demonstration of how to
use the ROS 2 backend to generate a minimal application would be a good
starting point for users that want to generate monitors for ROS 2.

This commit adds a markdown file under the `examples` directory that
explains how to use the ROS 2 backend to generate a simple monitoring
application.