ogma-core: Add field information to message in cFS. Refs #296.
The cFS template doesn't update the value of a variable correctly when
the value comes from a field of a cFS message.

Part of the problem is that the information needed in the cFS template
is not being made available, so there's no mechanism for the current cFS
template to know how to populate values from message fields.

This commit updates the cFS backend so that the information about the
message field, when applicable, is made available in the application
data that the cFS template can expand variables from.