ogma-core: Fix handling of message fields in cFS template. Refs #296.
The cFS template doesn't update the value of a variable correctly when
the value comes from a field of a cFS message.

A prior commit has modified the information made available to the
template, so that field names and types are available for all relevant
messages.

This commit updates the template to populate global variables from
struct fields when applicable.