ogma-core: Comment out include of monitor's implementation in cFS app. Refs #297.
The current cFS template mishandles how the Copilot code has to be
linked with the rest of the application. The CMakeLists.txt file
indicates that all C files in the application code directory must be
used. Before Copilot runs, the C files from the spec are not generated,
so CMake doesn't detect them. This is true the first time CMake runs,
but not afterwards. The C file for copilot.c is currently being included
using a C include, which conflicts with the way that CMake compiles and
links applications after the first time. As a consequence, the linked
may complain that step is being defined twice.

A prior commit has modified the CMakeLists.txt to explicitly list the
Copilot-generated C sources as dependencies to be compiled and linked
with the rest of the cFS application generated by Ogma.

This commit removes the inclusion of `copilot.c` into the cFS
application's main code, to prevent the same functions (e.g., `step`)
from being included twice.