ogma-cli: Add guide explaining basics of cFS backend. Refs #305.
To facilitate adoption of Ogma, we want to have a collection of examples
that demonstrate the most common use cases. A demonstration of how to
use the cFS backend to generate a minimal application would be a good
starting point for users that want to generate monitors for cFS.

This commit adds a markdown file under the `examples` directory that
explains how to use the cFS backend to generate a simple monitoring
application.