ogma-cli: Add example files to generate cFS monitoring application. Refs #305.
To facilitate adoption of Ogma, we want to have a collection of examples
that demonstrate the most common use cases. A demonstration of how to
use the cFS backend to generate a minimal application would be a good
starting point for users that want to generate monitors for cFS.

A prior commit added a guide explaining how to use the cFS backend to
generate a simple monitoring application.

This commit adds files that can be used with Ogma to generate a simple
monitoring application. The sample files match the explanation in the
aforementioned guide.