ogma-core: Avoid unnecessary recompilation of generated cFS app. Refs #299.
Compiling the Copilot code as part of the cFS build creates a sandbox (even if
one exists), calls cabal update (even if recently ran), and tries to reinstall
copilot (even if already installed). This additional overhead makes it
impossible to point to a sandbox in a different location, and also makes
compilation slow.

This commit updates the compilation rules in the `CMakeLists.txt` in the cFS
template so that `copilot` is only installed in the sandbox does not exist, and
so that the Copilot compiler is only executed in the source `Properties.hs`
file changes. This prevents unnecessary recompilation and allows users to
manually create a sandbox pointing elsewhere in their machines.