ogma-core: Adjust name of Copilot's step function name in cFS app. Refs #303.
The default cFS template generates a Copilot spec using the default step
function. Whenever the Copilot application gets a message and tries to
process it, it triggers a segmentation fault.

A prior commit modified the Copilot spec so that it produces C code that
uses the name `copilot_step` as entry point to re-evaluate the monitors.

This commit adjusts the cFS app so that it calls Copilot's entry point
by the new name.