ogma-core: Use alternate name for Copilot's step function in cFS template. Refs #303.
The default cFS template generates a Copilot spec using the default step
function. Whenever the Copilot application gets a message and tries to
process it, it triggers a segmentation fault.

This commit modifies the Copilot spec so that it produces C code that
uses the name `copilot_step` as entry point to re-evaluate the monitors.