Add sample image of runnning generated cFS application.
To facilitate adoption of Ogma, we want to have a collection of examples
that demonstrate the most common use cases. A demonstration of how to
use the cFS backend to generate a minimal application would be a good
starting point for users that want to generate monitors for cFS.

This commit adds an image that illustrates an Ogma-generated cFS
application begin executed and tested.