ogma-cli: Add CI job to test cFS backend. Refs #307.
To keep Ogma working in future versions and ensure that the changes
introduced do not affect the code generated for cFS in a negative way,
we want to have a test that checks that the code generated at least
compiles in a standard deployment of cFS.

This commit add a new CI action file that generates a cFS application,
adds it to a standard cFS bundle, enables the app, runs it, and sends
messages to it. The change does not check whether the app has received
the message or is reacting correctly, but at least checks that the cFS
instance doesn't crash when it receives messages.

This action is based on the first example available for cFS under
ogma-cli/examples.