ogma-language-c: Update license name, file in Cabal package. Refs #293.
Ogma has been approved for release under Apache license. We need to
update all license files, source code files, READMEs and cabal files
accordingly.

A prior commit has replaced the license file, which contained the NOSA
license, with a copy of the standard Apache version 2.0 license.

This commit updates the Cabal package definition to indicate that the
package has Apache version 2.0 license, and to point to the new license
file name in the same directory.