ogma-cli: Update license file in README. Refs #293.
Ogma has been approved for release under Apache license. We need to
update all license files, source code files, READMEs and cabal files
accordingly.

A prior commit has replaced the license file with a different file
with the same name, but without the extension.

This commit updates the license filed pointed to in the README so that
it refers to the proper filename in the repo.