ogma-cli: Add guide explaining basics of diagram backend. Refs #323.
To facilitate the adoption of Ogma, we want to have a collection of
examples that demonstrate the most common use cases. A demonstration of
how to use the diagram backend to generate a minimal state machine
implementation would be a good starting point for users that want to
generate monitors or applications from diagrams.

This commit adds a markdown file under the `examples` directory that
explains how to use the diagram backend to generate a state machine
implementation.