ogma-cli: Add example files to generate state machine. Refs #323.
To facilitate the adoption of Ogma, we want to have a collection of
examples that demonstrate the most common use cases. A demonstration of
how to use the diagram backend to generate a minimal state machine
implementation would be a good starting point for users that want to
generate monitors or applications from diagrams.

A prior commit added a guide explaining how to use the diagram backend
to generate a simple state machine.

This commit add a diagram and a main C program that can be used with
Ogma to generate a state machine. The sample files match the explanation
in the aforementioned guide.