ogma-core: Make state machine code return the current state. Refs #317.
The diagram to state machine generator produces an implementation of
state machines that returns the last state, rather than the current
state. The current behavior is due to the definition of `state` in
ogma/ogma-core/templates/diagram/Copilot.hs pre-inserting `initialState`
in the stream that is returned.

The current implementation makes the use of the code harder from the
integration point of view, since, normally, the current state is
desired.

This commit adjusts the template used by the diagram code generator to
define an additional `previousState` variable that has the initial state
pre-inserted, but return the current state at the end of the
computation.