ogma-core: Negate condition in state machine monitor. Refs #318.
The current diagram to state machine generator notifies with a trigger
if the current state machines the expectation, rather than the opposite.
As the most common use case of that checker is to detect invalid
transitions, rather than valid ones, we want to modify the condition to
check if the current state differs the expectation.

This commit modifies the diagram to state machine backend to use the
negation of the existing condition (that is, to use the condition that
the current and the expected state *do not match*) as trigger condition.