ogma-language-xlsx: Bump upper version constraint on xlsx. Refs #313.
In order to keep Ogma effectively working in the current Haskell
ecosystem, as well as new versions of GHC, we need to extend the
versions of dependencies that Ogma can be installed with. xlsx has seen
release 1.2.0, but constraints ih the cabal files prevent installation
with that version. This is preventing new versions of Ogma from being
added to the latest versions of Stackage.

This commit bumps the version constraints on xlsx so that Ogma can be
installed with versions of xlsx lower than 1.3.