ogma-cli: Add missing dependency to installation instructions in README. Refs #314.
The installation instructions are missing the dependency on `libz-dev`,
which makes it hard for people to follow them to install Ogma,
especially if they are not familiar with Haskell, where it is a common
requirement.

This commit modifies the installation instructions in the README to adds
`libz-dev` to the list of dependencies needed in the call to `apt-get`.